==> notebooks/1-Data Preprocessing.ipynb <==
!pip install -U 'spacy[cuda-autodetect]'
!pip install tqdm
!pip install contractions

import nltk
from nltk.corpus import stopwords
from nltk.cluster.util import cosine_distance
from nltk.tokenize import sent_tokenize, word_tokenize
import numpy as np
import pandas as pd
import networkx as nx
import re

from tqdm import tqdm
from tqdm._tqdm_notebook import tqdm_notebook
tqdm_notebook.pandas()

import spacy
gpu = spacy.prefer_gpu()
print('GPU:', gpu)

import contractions

train_df = pd.read_csv("./raw-data/train.csv")


## load the full dataset of 300k articles
train_df.head()

train_df.rename(columns={'highlights':'summary'}, inplace=True)

df = train_df.sample(n=22000)



# Initialize spacy 'en' model, keeping only tagger component needed for lemmatization
nlp = spacy.load('en_core_web_sm', disable=['parser', 'ner'])


def lemmatize_text(t):
    t = re.sub(r"\s+", " ", t)
    ### slang
    t = contractions.fix(t)


    # Parse the sentence using the loaded 'en' model object `nlp`
    doc = nlp(t)
    # Extract the lemma for each token and join
    return " ".join([token.lemma_ for token in doc])

df['article_lemmatized'] = df['article'].progress_apply(lemmatize_text)
df['summary_lemmatized'] = df['summary'].progress_apply(lemmatize_text)

df.head(1)

nltk.download('stopwords')
STOPWORDS = stopwords.words("english")
## add words that are too frequent
STOPWORDS = STOPWORDS + ["cnn","say","said","new","all", "due", "to", "on", "daily"]
## cleaning function
def clean_text(t, stopwords=None):
    ### separate sentences with '. '
    t = re.sub(r'\.(?=[^ \W\d])', '. ', str(t))
    ### remove punctuations and characters
    t = re.sub(r'[^\w\s]', '', t) 
    ### strip
    t = " ".join([word.strip() for word in t.split()])
    ### lowercase
    t = t.lower()

    ### tokenize (convert from string to list)
    t_list = t.split()
    
    ### remove Stopwords
    if stopwords is not None:
      t_list = [word for word in t_list if word not in 
                    stopwords]
    ### back to string
    t = " ".join(t_list)
    return t



df['article_cleaned'] = df['article_lemmatized'].progress_apply(clean_text, stopwords=STOPWORDS)
df['summary_cleaned'] = df['summary_lemmatized'].progress_apply(clean_text, stopwords=STOPWORDS)

df.head(1)

from sklearn.model_selection import train_test_split


train_df, test_df = train_test_split(df, test_size=2000)
train_df, val_df = train_test_split(train_df, test_size=4000)
print(len(train_df))
print(len(val_df))
print(len(test_df))

train_df.to_csv('./processed-data/train.csv')
val_df.to_csv('./processed-data/val.csv')
test_df.to_csv('./processed-data/test.csv')
