==> notebooks/4-Sentimental Analysis.ipynb <==
import numpy as np
from keras.datasets import imdb
from keras.models import Sequential
from keras.wrappers.scikit_learn import KerasClassifier
from keras.optimizers import Adam
from keras.layers import Dense, LSTM, Embedding
from keras.preprocessing.text import Tokenizer
from tensorflow.keras.preprocessing import sequence
from sklearn.model_selection import GridSearchCV

**Sentimental Analysis with LSTM**

max_features = 30000
maxlen=100

(X_train, y_train), (X_test, y_test) = imdb.load_data(num_words=max_features)
X_train = sequence.pad_sequences(X_train, maxlen=100)
X_test = sequence.pad_sequences(X_test, maxlen=100)

def build_model(lstm_units, dropout_rate, recurrent_dropout_rate, learning_rate):
    model = Sequential()
    model.add(Embedding(max_features, 128, input_length=maxlen))
    model.add(LSTM(lstm_units, dropout=dropout_rate, recurrent_dropout=recurrent_dropout_rate))
    model.add(Dense(1, activation='sigmoid'))
    model.compile(loss='binary_crossentropy', optimizer='adam', metrics=['accuracy'])
    return model

param_grid = {
    'lstm_units': [64, 128],
    'dropout_rate': [0.2, 0.3],
    'recurrent_dropout_rate': [0.1, 0.2],
    'learning_rate': [0.001, 0.005, 0.01],
}

model = KerasClassifier(build_fn=build_model, verbose=0)

grid = GridSearchCV(estimator=model, param_grid=param_grid, cv=3, verbose=1)
grid_result = grid.fit(X_test, y_test)

print(grid_result.best_params_)

# model1 = build_model(64, 0.3, 0.1, 0.005)
# model1.fit(X_train, y_train, batch_size=32, epochs=10, validation_data=(X_test, y_test))

# score, acc = model.evaluate(X_test, y_test, batch_size=32)
# print('Test score:', score)
# print('Test accuracy:', acc)

# model2 = build_model(64, 0.3, 0.1, 0.005)
# model2.fit(X_train, y_train, batch_size=32, epochs=6, validation_data=(X_test, y_test))
# score, acc = model2.evaluate(X_test, y_test, batch_size=32)
# print('Test score:', score)
# print('Test accuracy:', acc)

# model3 = build_model(64, 0.3, 0.1, 0.005)
# model3.fit(X_train, y_train, batch_size=64, epochs=6, validation_data=(X_test, y_test))
# score, acc = model3.evaluate(X_test, y_test, batch_size=64)
# print('Test score:', score)
# print('Test accuracy:', acc)

model = build_model(64, 0.3, 0.1, 0.005)
model.fit(X_train, y_train, batch_size=64, epochs=10, validation_data=(X_test, y_test))
score, acc = model.evaluate(X_test, y_test, batch_size=64)
print('Test score:', score)
print('Test accuracy:', acc)

model.summary()

try:
    from google.colab import drive
    drive.mount('/content/drive')
except ImportError:
    pass

import os
def get_root_dir():
    if os.path.exists('/content/drive/MyDrive/BT4222-Project'):
        return '/content/drive/MyDrive/BT4222-Project'
    else:
        return './'

print(get_root_dir())
os.chdir(get_root_dir())

import pandas as pd

val_df = pd.read_csv('/content/drive/MyDrive/BT4222-Project/processed-data/val.csv')
summary = val_df['summary'][:5]

tk = Tokenizer(num_words=100)
tk.fit_on_texts(summary)

sq = tk.texts_to_sequences(summary)
pad_sq = sequence.pad_sequences(sq, maxlen=100)

model.predict(pad_sq)

print(summary[0], '\n')
print(summary[1], '\n')
print(summary[2], '\n')
print(summary[3], '\n')
print(summary[4], '\n')
