==> notebooks/5-Text Summarization (Extractive).ipynb <==
import nltk
from nltk.corpus import stopwords
from nltk.cluster.util import cosine_distance
from nltk.tokenize import sent_tokenize
import numpy as np
import pandas as pd
import networkx as nx
import re

train_df = pd.read_csv('./processed-data/train.csv')
test_df = pd.read_csv('./processed-data/test.csv')
val_df = pd.read_csv('./processed-data/val.csv')

train_df.head()

# get stopwords
nltk.download('stopwords')
nltk.download('punkt')
STOPWORDS = stopwords.words('english')  
STOPWORDS

def get_sentences(text):        
  sentences = sent_tokenize(text)    
  for sentence in sentences:        
    sentence.replace("[^a-zA-Z0-9]"," ")     
  return sentences


def sentence_similarity(sent1,sent2,stopwords=[]):    
  sent1 = [w.lower() for w in sent1]    
  sent2 = [w.lower() for w in sent2]
        
  all_words = list(set(sent1 + sent2))   
     
  vector1 = [0] * len(all_words)    
  vector2 = [0] * len(all_words)        
  #build the vector for the first sentence    
  for w in sent1:        
    if not w in stopwords:
      vector1[all_words.index(w)]+=1                                                             
  #build the vector for the second sentence    
  for w in sent2:        
    if not w in stopwords:            
      vector2[all_words.index(w)]+=1 
               
  return 1-cosine_distance(vector1,vector2)


def build_similarity_matrix(sentences,stop_words):
  #create an empty similarity matrix
  similarity_matrix = np.zeros((len(sentences),len(sentences)))
  for idx1 in range(len(sentences)):
      for idx2 in range(len(sentences)):
        if idx1!=idx2:
          similarity_matrix[idx1][idx2] = sentence_similarity(sentences[idx1],sentences[idx2],stop_words)
  return similarity_matrix


@np.vectorize
def generate_summary(text,top_n):
  
  filtered_sentences=[]
  # Step1: read text and tokenize    
  sentences = get_sentences(text)
  # Step2: generate similarity matrix            
  sentence_similarity_matrix = build_similarity_matrix(sentences,STOPWORDS)
  # Step3: Rank sentences in similarity matrix
  sentence_similarity_graph = nx.from_numpy_array(sentence_similarity_matrix)
  scores = nx.pagerank(sentence_similarity_graph)
  # Step4: sort the rank and place top sentences
  ranked_sentences = sorted(((scores[i],s,i) for i,s in enumerate(sentences)),reverse=True)
  
  # Step5: get the top n number of sentences based on rank
  for i in range(top_n):
    filtered_sentences.append((ranked_sentences[i][1],ranked_sentences[i][2]))
  filtered_sentences.sort(key = lambda a: a[1])
  summarize_text = [s[0] for s in filtered_sentences]
  
  # Step6 : output the summarized version
  return " ".join(summarize_text)


#first 100 articles
articles = train_df.head(100)['article'].to_numpy()
articles[:5]

summaries = generate_summary(articles,3)

print('original text')

display(articles[1])
print()

print('summarized text')
display(summaries[1])

print('original text')

display(articles[50])
print()

print('summarized text')
display(summaries[50])
